Updated to newest yargs, abstract commands
