+++ package.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/openfn/core.git"
   },
   "main": "./lib/index.js",
-  "bin": "./lib/cli.js",
+  "bin": "./lib/cli/index.js",
   "scripts": {
     "test": "mocha --recursive",
     "test:watch": "mocha -w --recursive"
@@ -18,16 +18,15 @@
   "author": "Open Function Group",
   "license": "LGPL-3.0",
   "dependencies": {
-    "acorn": "^7.1.1",
-    "acorn-loose": "^7.0.0",
-    "ast-types": "^0.9.0",
+    "acorn": "^8.0.5",
+    "ast-types": "^0.14.2",
     "doclet-query": "github:openfn/doclet-query#v0.1.0",
     "estemplate": "^0.5.1",
     "fast-safe-stringify": "^2.0.7",
     "mitm": "^1.3.2",
     "recast": "0.19.0",
     "vm2": "^3.9.2",
-    "yargs": "^3.30.0"
+    "yargs": "^16.2.0"
   },
   "engines": {
     "node": ">= 12.17.0"
